requirements.txt
#-e git+git://github.com/citation-file-to_cff_string/cff-converter-python.git#egg=cff-converter-python
PyGitHub
pyapi-gitlab
